Added base callback for components and the advanced flow
